+++ src/util/environment.rs
@@ -286,8 +286,12 @@ pub const ORBIT_IP_CHECKSUM: &str = "ORBIT_IP_CHECKSUM";
 
 pub const ORBIT_TARGET: &str = "ORBIT_TARGET";
 pub const ORBIT_TOP_NAME: &str = "ORBIT_TOP_NAME";
+pub const ORBIT_TOP_FILE: &str = "ORBIT_TOP_FILE";
 pub const ORBIT_TB_NAME: &str = "ORBIT_TB_NAME";
+pub const ORBIT_TB_FILE: &str = "ORBIT_TB_FILE";
 pub const ORBIT_DUT_NAME: &str = "ORBIT_DUT_NAME";
+pub const ORBIT_DUT_FILE: &str = "ORBIT_DUT_FILE";
+
 pub const ORBIT_BLUEPRINT: &str = "ORBIT_BLUEPRINT";
 pub const ORBIT_TARGET_DIR: &str = "ORBIT_TARGET_DIR";
 pub const ORBIT_OUT_DIR: &str = "ORBIT_OUT_DIR";